+++ lib/src/conflicts.rs
@@ -65,17 +65,17 @@ fn write_diff_hunks(hunks: &[DiffHunk], file: &mut dyn Write) -> io::Result<()> 
                 debug_assert!(hunk.contents.iter().all_equal());
                 for line in hunk.contents[0].lines_with_terminator() {
                     file.write_all(b" ")?;
-                    file.write_all(line)?;
+                    write_and_ensure_newline(file, line)?;
                 }
             }
             DiffHunkKind::Different => {
                 for line in hunk.contents[0].lines_with_terminator() {
                     file.write_all(b"-")?;
-                    file.write_all(line)?;
+                    write_and_ensure_newline(file, line)?;
                 }
                 for line in hunk.contents[1].lines_with_terminator() {
                     file.write_all(b"+")?;
-                    file.write_all(line)?;
+                    write_and_ensure_newline(file, line)?;
                 }
             }
         }
@@ -462,14 +462,16 @@ fn materialize_git_style_conflict(
         conflict_marker_len,
         &format!("Side #1 ({conflict_info})"),
     )?;
-    output.write_all(left)?;
+    write_and_ensure_newline(output, left)?;
+
     write_conflict_marker(
         output,
         ConflictMarkerLineChar::GitAncestor,
         conflict_marker_len,
         "Base",
     )?;
-    output.write_all(base)?;
+    write_and_ensure_newline(output, base)?;
+
     // VS Code doesn't seem to support any trailing text on the separator line
     write_conflict_marker(
         output,
@@ -477,7 +479,8 @@ fn materialize_git_style_conflict(
         conflict_marker_len,
         "",
     )?;
-    output.write_all(right)?;
+
+    write_and_ensure_newline(output, right)?;
     write_conflict_marker(
         output,
         ConflictMarkerLineChar::ConflictEnd,
@@ -503,7 +506,7 @@ fn materialize_jj_style_conflict(
             conflict_marker_len,
             &format!("Contents of side #{}", add_index + 1),
         )?;
-        output.write_all(data)
+        write_and_ensure_newline(output, data)
     };
 
     // Write a negative snapshot (base) of a conflict
@@ -514,7 +517,7 @@ fn materialize_jj_style_conflict(
             conflict_marker_len,
             &format!("Contents of {base_str}"),
         )?;
-        output.write_all(data)
+        write_and_ensure_newline(output, data)
     };
 
     // Write a diff from a negative term to a positive term
@@ -591,6 +594,21 @@ fn materialize_jj_style_conflict(
         &format!("{conflict_info} ends"),
     )?;
     Ok(())
+}
+
+// Write a chunk of data, ensuring that it doesn't end with a line which is
+// missing its terminating newline.
+fn write_and_ensure_newline(output: &mut dyn Write, data: &[u8]) -> io::Result<()> {
+    output.write_all(data)?;
+    if has_no_eol(data) {
+        writeln!(output)?;
+    }
+    Ok(())
+}
+
+// Check whether a slice is missing its terminating newline character.
+fn has_no_eol(slice: &[u8]) -> bool {
+    slice.last().is_some_and(|&last| last != b'\n')
 }
 
 fn diff_size(hunks: &[DiffHunk]) -> usize {
@@ -870,7 +888,7 @@ pub async fn update_from_content(
 
     // Parse conflicts from the new content using the arity of the simplified
     // conflicts.
-    let Some(hunks) = parse_conflict(
+    let Some(mut hunks) = parse_conflict(
         content,
         simplified_file_ids.num_sides(),
         conflict_marker_len,
@@ -879,6 +897,17 @@ pub async fn update_from_content(
         let file_id = store.write_file(path, &mut &content[..]).await?;
         return Ok(Merge::normal(file_id));
     };
+
+    // If there is a conflict at the end of the file and a term ends with a newline,
+    // check whether the original term ended with a newline. If it didn't, then
+    // remove the newline since it was added automatically when materializing.
+    if let Some(last_hunk) = hunks.last_mut().filter(|hunk| !hunk.is_resolved()) {
+        for (original_content, term) in merge_hunk.iter().zip_eq(last_hunk.iter_mut()) {
+            if term.last() == Some(&b'\n') && has_no_eol(original_content) {
+                term.pop();
+            }
+        }
+    }
 
     let mut contents = simplified_file_ids.map(|_| vec![]);
     for hunk in hunks {